Add files via upload
A snake is controlled by the player as it walks around a rectangular board in the Python-based Snake Game project, dodging obstacles like walls and its own tail while consuming food to become longer. The Pygame module in Python is used to develop the game since it has the resources needed to build a graphical user interface and manage user input. The aim of the game is to consume as much food as you can while avoiding the obstacles in order to earn the best score. The game may be made even more unique by including elements like sound effects, several difficulty settings, and high-score monitoring.